Initial refactor to break up CoreGenerator
